rename a header extension from .h to .hpp
