+++ src/logger.cpp
@@ -1,4 +1,4 @@
-#include "logger.h"
+#include "logger.hpp"
 
 
 LogLevel Logger::m_msgLevel = LogLevel::info;
