+++ src/space.hpp
@@ -1,7 +1,7 @@
 #pragma once
 
 #include <list>
-#include "game_entity.h"
+#include "game_entity.hpp"
 
 class Space
 {
