+++ tests/game_entity_test.cpp
@@ -1,9 +1,9 @@
 #include "gtest/gtest.h"
-#include "alien.h"
-#include "space_ship.h"
-#include "obstacle.h"
-#include "bullet_manager.h"
-#include "space.h"
+#include "alien.hpp"
+#include "space_ship.hpp"
+#include "obstacle.hpp"
+#include "bullet_manager.hpp"
+#include "space.hpp"
 #include "except.hpp"
 
 TEST(bullet_test, test_construction)
